code smell: dead code - owned_books, data clumps
